+++ src/eps/interface/epsregis.c
@@ -49,6 +49,9 @@ SLEPC_EXTERN PetscErrorCode EPSCreate_ELPA(EPS);
 #if defined(SLEPC_HAVE_ELEMENTAL)
 SLEPC_EXTERN PetscErrorCode EPSCreate_Elemental(EPS);
 #endif
+#if defined(SLEPC_HAVE_EVSL)
+SLEPC_EXTERN PetscErrorCode EPSCreate_EVSL(EPS);
+#endif
 
 /*@C
   EPSRegisterAll - Registers all the eigenvalue solvers in the EPS package.
@@ -104,6 +107,9 @@ PetscErrorCode EPSRegisterAll(void)
 #endif
 #if defined(SLEPC_HAVE_ELEMENTAL)
   ierr = EPSRegister(EPSELEMENTAL,EPSCreate_Elemental);CHKERRQ(ierr);
+#endif
+#if defined(SLEPC_HAVE_EVSL)
+  ierr = EPSRegister(EPSEVSL,EPSCreate_EVSL);CHKERRQ(ierr);
 #endif
   PetscFunctionReturn(0);
 }